+++ src/components/ui/generalNote/theme.js
@@ -3,5 +3,10 @@ export const generalNoteTheme = {
        success: 'green',
        warning: 'orange',
        error: 'red'    
+    },
+    blue: {
+        success: 'blue',
+        warning: 'orange',
+        error: 'red' 
     }
 };