+++ src/index.js
@@ -1,6 +1,5 @@
 import './style.css';
 import './styles/index';
-import './components/common'
 import './theme/index';
 import './enums';
 import './settings';