+++ src/settings/sets/events.js
@@ -1,5 +1,6 @@
 export const CommonEvents = {
     click: 'click',
+    tap: 'tap',
     resize: 'resize',
     keypress: 'keypress',
     keydown: 'keydown',