+++ src/theme/enums.js
@@ -1,0 +1,7 @@
+export const Themes = {
+    main1: 'main1',
+    blue: 'blue',
+    black: 'black',
+    red: 'red',
+    yellow: 'yellow'
+};